+++ systems/input_system.cpp
@@ -14,7 +14,7 @@ void InputSystem::process_input(Globals& globals, std::list<std::pair<EntityID, 
         switch(m_event.type)
         {
             case SDL_QUIT:
-            	globals(Globals::app_running) = CommandReturnValue{0l};
+            	globals(Globals::app_running) = CommandReturnValue{0.0};
             break;
 
             case SDL_KEYDOWN:
@@ -37,9 +37,9 @@ void InputSystem::process_input(Globals& globals, std::list<std::pair<EntityID, 
                 else if(m_event.key.keysym.sym == m_keyselect)
                     m_select_pressed = true;
                 else if(m_event.key.keysym.sym == m_keyquit || m_event.key.keysym.sym == SDLK_F10)
-                	globals(Globals::app_running) = CommandReturnValue{0l};
+                	globals(Globals::app_running) = CommandReturnValue{0.0};
                 else if(m_event.key.keysym.sym == SDLK_r)
-                	globals(Globals::app_needs_reload) = CommandReturnValue{1l};
+                	globals(Globals::app_needs_reload) = CommandReturnValue{static_cast<int64_t>(1)};
                 else if(m_event.key.keysym.sym == m_keypause)
                 	globals(Globals::app_paused) = CommandReturnValue{!globals(Globals::app_paused).boolean()};
                 else if(m_event.key.keysym.sym == SDLK_h)
